Add storybook test for popover, remove popoverOpen as it was only needed for unused storybook test
