+++ client/web/src/search/results/streaming-progress/StreamingProgress.story.tsx
@@ -110,47 +110,3 @@ add('2 results from 2 repositories, complete, skipped with warning', () => {
 
     return <WebStory>{() => <StreamingProgress progress={progress} />}</WebStory>
 })
-
-add('2 results from 2 repositories, complete, skipped with warning, popover open', () => {
-    const progress: Progress = {
-        done: true,
-        durationMs: 1500,
-        matchCount: 2,
-        repositoriesCount: 2,
-        skipped: [
-            {
-                reason: 'excluded-fork',
-                message: '',
-                severity: 'info',
-                title: '10k forked repositories excluded',
-                suggested: {
-                    title: 'include forked',
-                    queryExpression: 'fork:yes',
-                },
-            },
-            {
-                reason: 'excluded-archive',
-                message: '',
-                severity: 'info',
-                title: '60k archived repositories excluded',
-                suggested: {
-                    title: 'include archived',
-                    queryExpression: 'archived:yes',
-                },
-            },
-            {
-                reason: 'shard-timedout',
-                message:
-                    'Search timed out before some repositories could be searched. Try reducing scope of your query with repo: or other filters.',
-                severity: 'warn',
-                title: 'Search timed out',
-                suggested: {
-                    title: 'increase timeout',
-                    queryExpression: 'timeout:2m',
-                },
-            },
-        ],
-    }
-
-    return <WebStory>{() => <StreamingProgress progress={progress} popoverOpen={true} />}</WebStory>
-})
