+++ client/web/src/search/results/streaming-progress/StreamingProgress.tsx
@@ -5,7 +5,6 @@ import { StreamingProgressSkippedButton } from './StreamingProgressSkippedButton
 
 export interface StreamingProgressProps {
     progress?: Progress
-    popoverOpen?: boolean
     onSearchAgain?: (additionalFilters: string[]) => void
 }
